+++ Gym.java
@@ -1,7 +1,7 @@
 public class Gym extends Amenity {
     public Gym(){};
 
-    public Gym(String name, String description, int price){
+    public Gym(String name, String description, double price){
         super(name, description, price);
     }
 }
